% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_helpers.R
\name{create_lower_extension}
\alias{create_lower_extension}
\title{Computes the lower boundary value for the empirical CDF}
\usage{
create_lower_extension(x)
}
\arguments{
\item{x, }{a numeric vector}
}
\value{
numeric
}
\description{
Computes the lower boundary value for the empirical CDF
}
\keyword{internal}
