% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_helpers.R
\name{create_upper_extension}
\alias{create_upper_extension}
\title{Computes the upper boundary value for the empirical CDF}
\usage{
create_upper_extension(x)
}
\arguments{
\item{x, }{a numeric vector}
}
\value{
numeric
}
\description{
Computes the upper boundary value for the empirical CDF
}
\keyword{internal}
