% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf.R
\name{demp}
\alias{demp}
\title{Empirical probability density function (EPDF)}
\usage{
demp(x, sample)
}
\arguments{
\item{x}{numeric vector of values to evaluate}

\item{sample}{numeric vector of sample values to base the EPDF on}
}
\value{
numeric vector of density values based on the EPDF
}
\description{
Empirical probability density function (EPDF)
}
\examples{
x <- 1:5
demp(1, x)
}
