% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{estimate_one_n}
\alias{estimate_one_n}
\title{Estimate N}
\usage{
estimate_one_n(sample, alpha, power, delta)
}
\arguments{
\item{sample}{pilot data}

\item{alpha}{Type I error probability}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}
}
\value{
numeric value, an estimate of the sample size required to
detect a location shift of \code{delta} with a Wilcoxon Mann-Whitney
test with \code{power} and \code{alpha}.
}
\description{
Estimate N
}
\examples{
estimate_one_n(sample = 1:5, alpha = 0.05, power = 0.9, delta = 0.5)
}
