% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_helpers.R
\name{extend_sample}
\alias{extend_sample}
\title{Extends a vector by adding a lower and upper boundary.}
\usage{
extend_sample(x)
}
\arguments{
\item{x, }{a numeric vector}
}
\value{
extended and sorted numeric vector
}
\description{
Extends a vector by adding a lower and upper boundary.
}
\keyword{internal}
