% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf_helpers.R
\name{find_smaller_index}
\alias{find_smaller_index}
\title{Finds the index of the smaller neighbour of the given value in the vector x.}
\usage{
find_smaller_index(value, x)
}
\arguments{
\item{value, }{value whose neighbour is searched}

\item{x, }{numeric vector}
}
\value{
integer
}
\description{
Finds the index of the smaller neighbour of the given value in the vector x.
}
\keyword{internal}
