% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cundill.R
\name{n_gamma}
\alias{n_gamma}
\title{Calculate sample size for gamma distribution}
\usage{
n_gamma(mean0, shape, efficacy, alpha = 0.05, power = 0.9, q = 0.5, link = log)
}
\arguments{
\item{mean0}{Mean in control group}

\item{shape}{Shape, assumed equal across groups}

\item{efficacy}{Effect size (1 - mean0/mean1)}

\item{alpha}{Type I error rate}

\item{power}{1 - Type II error rate}

\item{q}{Proportion of observations allocated to the control group}

\item{link}{Link function}
}
\value{
Required number of observations
}
\description{
Calculate sample size for gamma distribution
}
