% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cundill.R
\name{n_general}
\alias{n_general}
\title{The general formula for sample size calculation as published by
Cundill & Alexander (2015)}
\usage{
n_general(
  spec = list(mean0 = NULL, mean1 = NULL, q0 = NULL, q1 = NULL, dmu_eta = NULL,
    variance = NULL, linkfun = NULL)
)
}
\arguments{
\item{spec}{A list that holds the necessary specification.}
}
\value{
Numeric value, the required sample size
}
\description{
The general formula for sample size calculation as published by
Cundill & Alexander (2015)
}
