% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti.R
\name{n_locshift}
\alias{n_locshift}
\title{Estimate N on the basis of two pilot samples.}
\usage{
n_locshift(s1, s2, delta, power, alpha)
}
\arguments{
\item{s1, s2}{Pilot samples}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{alpha}{Type I error probability}
}
\value{
numeric value, an estimate of the sample size required to
detect a location shift of \code{delta} with a Wilcoxon Mann-Whitney
test with \code{power} and \code{alpha}.
}
\description{
Estimate N on the basis of two pilot samples.
}
\examples{
n_locshift(s1 = rnorm(10), s2 = rnorm(10), alpha = 0.05, power = 0.9, delta = 0.5)
}
