% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cundill.R
\name{n_negbinom}
\alias{n_negbinom}
\title{Calculate sample size for negative binomial distribution}
\usage{
n_negbinom(
  mean0,
  efficacy,
  dispersion0,
  dispersion1 = dispersion0,
  alpha = 0.05,
  power = 0.9,
  q = 0.5,
  link_fun = log,
  dmu_deta_fun = function(mu) mu,
  two_sided = TRUE
)
}
\arguments{
\item{mean0}{Mean in control group}

\item{efficacy}{Effect size, \eqn{1 - (\mu_1 / \mu_0)}, where
\eqn{\mu_0} is the mean in the control group (\code{mean0}) and
\eqn{\mu_1} is the mean in the treatment group.}

\item{dispersion0}{Dispersion parameter in control group}

\item{dispersion1}{Dispersion parameter in treatment group. Defaults to
\code{shape0}, because GLM assumes equal shape across groups.}

\item{alpha}{Type I error rate}

\item{power}{1 - Type II error rate}

\item{q}{Proportion of observations allocated to the control group}

\item{link_fun}{function object, the link function to use}

\item{dmu_deta_fun}{function object, derivative of the original
mean with respect to the link: \eqn{d\mu / d\eta}. Must fit
the specified link function. In the default case of a log link,
we have \eqn{d\mu / d\eta = \mu}.}

\item{two_sided}{logical, if \code{TRUE} the sample size
will be calculated for a two-sided test. Otherwise, the sample
size will be calculated for a one-sided test.}
}
\value{
Returns an object of class \code{"sample_size"}. It contains
the following components:
\item{n}{the total sample size}
\item{n1}{sample size in Group 1 (control group)}
\item{n2}{sample size in Group 2 (treatment group)}
\item{two_sided}{logical, \code{TRUE}, if the estimated sample size
refers to a two-sided test}
\item{alpha}{type I error rate used in sample size estimation}
\item{power}{target power used in sample size estimation}
\item{effect}{effect size used in sample size estimation}
\item{effect_type}{short description of the type of effect size}
\item{comment}{additional comment, if there is any}
\item{call}{the matched call.}
}
\description{
Citation: Cundill & Alexander (2015)
}
\details{
Cundill, B., & Alexander, N. D. E. (2015). Sample size calculations
for skewed distributions. \emph{BMC Medical Research Methodology},
15(1), 1–9. https://doi.org/10.1186/s12874-015-0023-0
}
\examples{
n_negbinom(mean0 = 71.4, efficacy = 0.7, dispersion0 = 0.33,
           alpha = 0.05, power = 0.9)
}
