% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{n_noether}
\alias{n_noether}
\title{Noether's (1987) formula for obtaining a sample size estimation for
the Wilcoxon Mann-Whitney test.}
\usage{
n_noether(alpha, power, p)
}
\arguments{
\item{alpha}{Type I error probability}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{p}{probability \eqn{P(X < Y)} that a random observation from
group X is smaller than a random observation from group Y}
}
\value{
estimated required sample size
}
\description{
Noether's (1987) formula for obtaining a sample size estimation for
the Wilcoxon Mann-Whitney test.
}
