% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf.R
\name{pemp}
\alias{pemp}
\title{Empirical cumulative density function (ECDF)}
\usage{
pemp(q, sample)
}
\arguments{
\item{q}{numeric vector of values to evaluate}

\item{sample}{numeric vector of sample values to base the ECDF on}
}
\value{
Returns the probabilities that a value drawn at random from the
empirical cumulative density based on \emph{sample} is smaller than or
equal to the elements of x.
}
\description{
Empirical cumulative density function (ECDF)
}
\examples{
x <- 1:5
pemp(1, x)
}
