% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_cdf.R
\name{qemp}
\alias{qemp}
\title{Empirical quantile function}
\usage{
qemp(p, sample)
}
\arguments{
\item{p}{probability}

\item{sample}{numeric vector of sample values to base the ECDF on}
}
\value{
Returns the value for which \code{pemp(x, sample) = p},
i.e. the probability that a value drawn at random from the ECDF
is smaller or equal to \code{x} is \code{p}.
}
\description{
Empirical quantile function
}
\examples{
x <- 1:5
qemp(0.1, x)
}
