% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti.R
\name{resample_n_locshift}
\alias{resample_n_locshift}
\title{Compute \code{n_resamples} estimates of N based on two pilot samples.}
\usage{
resample_n_locshift(s1, s2, n_resamples, delta, power, alpha)
}
\arguments{
\item{s1, s2}{Pilot samples}

\item{n_resamples}{number of resamples to use in bootstrapping}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{alpha}{Type I error probability}
}
\value{
numeric vector of sample size estimates
}
\description{
Compute \code{n_resamples} estimates of N based on two pilot samples.
}
