% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{resample_n_locshift_one}
\alias{resample_n_locshift_one}
\title{Compute \code{n_resamples} estimates of N}
\usage{
resample_n_locshift_one(sample, n_resamples, alpha, power, delta)
}
\arguments{
\item{sample}{pilot data}

\item{n_resamples}{number of resamples to use in bootstrapping}

\item{alpha}{Type I error probability}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}
}
\value{
numeric vector of sample size estimates
}
\description{
Compute \code{n_resamples} estimates of N
}
