% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{resample_one_n}
\alias{resample_one_n}
\title{Compute \code{n_resamples} estimates of N}
\usage{
resample_one_n(sample, n_resamples, ...)
}
\arguments{
\item{sample}{pilot data}

\item{n_resamples}{number of resamples to use in bootstrapping}

\item{...}{arguments passed on to \code{\link{estimate_one_n}}.}
}
\value{
numeric vector of sample size estimates
}
\description{
Compute \code{n_resamples} estimates of N
}
