% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{estimate_p}
\alias{estimate_p}
\title{Computes an empirical estimate of \code{p} (\eqn{P(X < X + \delta)})}
\usage{
estimate_p(sample, delta)
}
\arguments{
\item{sample}{numeric vector of data to base the estimation on (X)}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}
}
\value{
An empirical estimate of \eqn{P(X < X + \delta)}
}
\description{
Computes an empirical estimate of \code{p} (\eqn{P(X < X + \delta)})
}
\keyword{internal}
