% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_chakraborti_helpers.R
\name{n_locshift_one}
\alias{n_locshift_one}
\title{Estimate N on the basis of one pilot sample.}
\usage{
n_locshift_one(sample, alpha, power, delta, q)
}
\arguments{
\item{sample}{pilot data}

\item{alpha}{Type I error probability}

\item{power}{1 - Type II error probability, the desired statistical
power}

\item{delta}{numeric value, location shift parameter \eqn{\delta}}

\item{q}{size of group0 relative to total sample size.}
}
\value{
numeric value, an estimate of the sample size required to
detect a location shift of \code{delta} with a Wilcoxon Mann-Whitney
test with \code{power} and \code{alpha}.
}
\description{
Estimate N on the basis of one pilot sample.
}
\keyword{internal}
